Basic setup for components based on designs provided
