+++ src/components/navigation.vue
@@ -1,0 +1,10 @@
+<template>
+  <div class="bt-main-nav">
+   <h1>MAIN NAV</h1>
+  </div>
+</template>
+
+<!-- Add "scoped" attribute to limit CSS to this component only -->
+<style scoped lang="scss">
+
+</style>
