+++ src/components/transaction-header.vue
@@ -1,0 +1,19 @@
+<template>
+  <div class="bt-transaction-header">
+  <h1> Transaction Header </h1>
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'TransactionHeader',
+  props: {
+    total: Number
+  }
+}
+</script>
+
+<!-- Add "scoped" attribute to limit CSS to this component only -->
+<style scoped lang="scss">
+
+</style>
