+++ src/components/transaction-row.vue
@@ -1,0 +1,18 @@
+<template>
+  <div class="bt-transaction-row">
+  </div>
+</template>
+
+<script>
+export default {
+  name: 'TransactionRow',
+  props: {
+    msg: String
+  }
+}
+</script>
+
+<!-- Add "scoped" attribute to limit CSS to this component only -->
+<style scoped lang="scss">
+
+</style>
